flask/requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
MarkupSafe==2.1.1
Flask-Babel==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.10.1
python-dotenv==1.0.1